Broke DLCMessage into many files
