+++ core/src/main/scala/org/bitcoins/core/protocol/dlc/OracleOutcome.scala
@@ -1,10 +1,5 @@
 package org.bitcoins.core.protocol.dlc
 
-import org.bitcoins.core.protocol.dlc.DLCMessage.{
-  EnumSingleOracleInfo,
-  NumericSingleOracleInfo,
-  SingleOracleInfo
-}
 import org.bitcoins.core.protocol.tlv.{
   DLCOutcomeType,
   EnumOutcome,
