+++ testkit/src/main/scala/org/bitcoins/testkit/core/gen/TLVGen.scala
@@ -1,14 +1,10 @@
 package org.bitcoins.testkit.core.gen
 
-import org.bitcoins.core.protocol.dlc.DLCMessage.{
-  ContractInfo,
-  DLCAccept,
-  DLCOffer
-}
+import org.bitcoins.core.protocol.dlc.DLCMessage.{DLCAccept, DLCOffer}
 import org.bitcoins.core.config.Networks
 import org.bitcoins.core.currency.{Bitcoins, CurrencyUnit, Satoshis}
 import org.bitcoins.core.number.UInt32
-import org.bitcoins.core.protocol.dlc.DLCFundingInputP2WPKHV0
+import org.bitcoins.core.protocol.dlc.{ContractInfo, DLCFundingInputP2WPKHV0}
 import org.bitcoins.core.protocol.tlv._
 import org.bitcoins.core.protocol.transaction._
 import org.bitcoins.core.protocol.{BigSizeUInt, BlockTimeStamp}
